new class and moving stuff around
