+++ src/polyfem/mesh/remesh/wild_remesh/LocalRelaxationData.cpp
@@ -4,8 +4,8 @@
 #include <polyfem/mesh/remesh/WildRemesher.hpp>
 #include <polyfem/assembler/Mass.hpp>
 #include <polyfem/solver/forms/ContactForm.hpp>
-#include <polyfem/solver/forms/BCLagrangianForm.hpp>
-#include <polyfem/solver/forms/BCPenaltyForm.hpp>
+#include <polyfem/solver/forms/lagrangian/BCLagrangianForm.hpp>
+#include <polyfem/solver/forms/lagrangian/BCPenaltyForm.hpp>
 #include <polyfem/solver/forms/ContactForm.hpp>
 #include <polyfem/solver/problems/StaticBoundaryNLProblem.hpp>
 #include <polyfem/time_integrator/ImplicitTimeIntegrator.hpp>
