+++ src/polyfem/solver/ALSolver.cpp
@@ -5,8 +5,8 @@
 namespace polyfem::solver
 {
 	ALSolver::ALSolver(
-		std::shared_ptr<BCLagrangianForm> lagr_form,
-		std::shared_ptr<BCPenaltyForm> pen_form,
+		std::shared_ptr<LagrangianForm> lagr_form,
+		std::shared_ptr<LagrangianPenaltyForm> pen_form,
 		const double initial_al_weight,
 		const double scaling,
 		const double max_al_weight,
