+++ src/polyfem/solver/SolveData.cpp
@@ -2,8 +2,8 @@
 
 #include <polyfem/solver/NLProblem.hpp>
 #include <polyfem/solver/forms/Form.hpp>
-#include <polyfem/solver/forms/BCLagrangianForm.hpp>
-#include <polyfem/solver/forms/BCPenaltyForm.hpp>
+#include <polyfem/solver/forms/lagrangian/BCLagrangianForm.hpp>
+#include <polyfem/solver/forms/lagrangian/BCPenaltyForm.hpp>
 #include <polyfem/solver/forms/MacroStrainLagrangianForm.hpp>
 #include <polyfem/solver/forms/MacroStrainALForm.hpp>
 #include <polyfem/solver/forms/BodyForm.hpp>
