+++ src/polyfem/solver/SolveData.hpp
@@ -38,8 +38,8 @@ namespace polyfem::solver
 	class MacroStrainALForm;
 	class FrictionForm;
 	class BodyForm;
-	class BCLagrangianForm;
-	class BCPenaltyForm;
+	class LagrangianForm;
+	class LagrangianPenaltyForm;
 	class MacroStrainLagrangianForm;
 	class MacroStrainALForm;
 	class InertiaForm;
@@ -139,8 +139,8 @@ namespace polyfem::solver
 		std::shared_ptr<assembler::PressureAssembler> pressure_assembler;
 		std::shared_ptr<solver::NLProblem> nl_problem;
 
-		std::shared_ptr<solver::BCLagrangianForm> al_lagr_form;
-		std::shared_ptr<solver::BCPenaltyForm> al_pen_form;
+		std::shared_ptr<solver::LagrangianForm> al_lagr_form;
+		std::shared_ptr<solver::LagrangianPenaltyForm> al_pen_form;
 		std::shared_ptr<solver::MacroStrainLagrangianForm> strain_al_lagr_form;
 		std::shared_ptr<solver::MacroStrainALForm> strain_al_pen_form;
 		std::shared_ptr<solver::BodyForm> body_form;
