+++ src/polyfem/solver/forms/lagrangian/CMakeLists.txt
@@ -1,0 +1,12 @@
+set(SOURCES
+	BCLagrangianForm.cpp
+	BCLagrangianForm.hpp
+	BCPenaltyForm.cpp
+	BCPenaltyForm.hpp
+	CMakeLists.txt
+	LagrangianForm.hpp
+	LagrangianPenaltyForm.hpp
+)
+
+source_group(TREE "${CMAKE_CURRENT_SOURCE_DIR}" PREFIX "Source Files" FILES ${SOURCES})
+target_sources(polyfem PRIVATE ${SOURCES})
